model: Moving materialized_topic type from v/coproc to v/model
- This allows other portions of the codebase to be materialized topic aware
without having to link against v/coproc
